requirements.txt
asgiref==3.3.4
backcall==0.2.0
beautifulsoup4==4.9.3
cent==3.0.1
certifi==2021.5.30
chardet==4.0.0
decorator==5.0.6
Django==2.1.15
django-bootstrap==0.2.4
django-bootstrap3==15.0.0
django-bootstrap4==3.0.1
gunicorn==20.1.0
idna==2.10
importlib-metadata==2.1.1
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.8.1
PyJWT==2.1.0
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
traitlets==4.3.3
typing-extensions==3.10.0.0
urllib3==1.26.5
wcwidth==0.2.5
zipp==3.4.1
